🚚 api: [SHARED] move adapter to shared context
